Move PageHeader and associated components into reusable UI directory
