+++ ui/src/flux/components/EmptyFluxPage.tsx
@@ -1,6 +1,6 @@
 import React, {SFC} from 'react'
 
-import PageHeader from 'src/shared/components/PageHeader'
+import PageHeader from 'src/reusable_ui/components/page_layout/PageHeader'
 
 interface Props {
   onShowOverlay: () => void
