+++ ui/test/admin/containers/AdminInfluxDBPage.test.tsx
@@ -2,8 +2,8 @@ import React from 'react'
 import {shallow} from 'enzyme'
 
 import {DisconnectedAdminInfluxDBPage} from 'src/admin/containers/AdminInfluxDBPage'
-import PageHeader from 'src/shared/components/PageHeader'
-import Title from 'src/shared/components/PageHeaderTitle'
+import PageHeader from 'src/reusable_ui/components/page_layout/PageHeader'
+import Title from 'src/reusable_ui/components/page_layout/PageHeaderTitle'
 import {source} from 'test/resources'
 
 describe('AdminInfluxDBPage', () => {
